+++ src/hb/aat_layout_kerx_table.rs
@@ -4,10 +4,10 @@ use ttf_parser::{ankr, apple_layout, kerx, FromData, GlyphId};
 
 use crate::hb::buffer::{hb_buffer_t, BufferScratchFlags};
 use crate::hb::hb_font_t;
-use crate::hb::ot::matching::SkippyIter;
-use crate::hb::ot::{lookup_flags, ApplyContext};
 use crate::hb::ot_layout::TableIndex;
+use crate::hb::ot_layout_common::lookup_flags;
 use crate::hb::ot_layout_gpos_table::attach_type;
+use crate::hb::ot_layout_gsubgpos::{ApplyContext, SkippyIter};
 use crate::hb::shape_plan::hb_ot_shape_plan_t;
 
 trait ExtendedStateTableExt<T: FromData + Copy> {
