+++ src/hb/mod.rs
@@ -7,12 +7,15 @@ mod aat_layout_trak_table;
 mod aat_map;
 pub mod common;
 pub mod face;
+mod feature;
 mod glyph_set;
+mod kerning;
 mod machine_cursor;
-mod ot;
 mod ot_layout;
+mod ot_layout_common;
 mod ot_layout_gpos_table;
 mod ot_layout_gsub_table;
+mod ot_layout_gsubgpos;
 mod ot_map;
 mod ot_shape_complex;
 mod ot_shape_complex_arabic;
